+++ src/store/features/assets/assetsState.ts
@@ -1,28 +1,29 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 
-interface CachedWarnings {
+type BackgroundAsset = {
   id: string;
   is360: boolean;
   warnings: string[];
   numTiles: number;
+  lookup: Uint8Array;
   timestamp: number;
-}
+};
 
-export interface WarningsState {
+export interface AssetsState {
   backgroundsLoading: boolean;
-  backgrounds: Record<string, CachedWarnings>;
+  backgrounds: Record<string, BackgroundAsset>;
 }
 
-export const initialState: WarningsState = {
+export const initialState: AssetsState = {
   backgroundsLoading: false,
   backgrounds: {},
 };
 
-const warningsSlice = createSlice({
-  name: "warnings",
+const assetsSlice = createSlice({
+  name: "assets",
   initialState,
   reducers: {
-    checkBackgroundWarnings: (
+    loadBackgroundAssetInfo: (
       state,
       _action: PayloadAction<{
         backgroundId: string;
@@ -31,13 +32,14 @@ const warningsSlice = createSlice({
     ) => {
       state.backgroundsLoading = true;
     },
-    setBackgroundWarnings: (
+    setBackgroundAssetInfo: (
       state,
       action: PayloadAction<{
         id: string;
         is360: boolean;
         warnings: string[];
         numTiles: number;
+        lookup: Uint8Array;
       }>
     ) => {
       state.backgroundsLoading = false;
@@ -46,12 +48,13 @@ const warningsSlice = createSlice({
         is360: action.payload.is360,
         warnings: action.payload.warnings,
         numTiles: action.payload.numTiles,
+        lookup: action.payload.lookup,
         timestamp: Date.now(),
       };
     },
   },
 });
 
-export const { actions, reducer } = warningsSlice;
+export const { actions, reducer } = assetsSlice;
 
 export default reducer;